update: common integrations test code
